Refactor credential helper and testing code out of common.
