+++ test/units/module_utils/docker/test_common.py
@@ -1,34 +1,14 @@
 from __future__ import (absolute_import, division, print_function)
 __metaclass__ = type
 
-import os
 import pytest
 
-from docker.constants import DEFAULT_DOCKER_API_VERSION
-
-from units.modules.utils import patch, set_module_args
-from ansible.module_utils import basic
-
 from ansible.module_utils.docker.common import (
-    Store,
-    StoreError,
-    DockerFileStore,
-    AnsibleDockerClient,
     compare_dict_allow_more_present,
     compare_generic,
     convert_duration_to_nanosecond,
     parse_healthcheck
 )
-
-curr_dir = os.path.dirname(__file__)
-test_data_dir = os.path.join(curr_dir, "test_data")
-
-EXAMPLE_REGISTRY = "registry.example.com"
-CREDENTIAL_HELPERS = [
-    'desktop-config.json',
-    'osxkeychain-config.json',
-    'secretservice-config.json'
-]
 
 DICT_ALLOW_MORE_PRESENT = (
     {
@@ -536,46 +516,3 @@ def test_parse_healthcheck():
         'interval': 3662003004000
     }
     assert disabled is False
-
-
-def get_docker_client():
-    '''
-    Avoid crash while instantiating AnsibleDockerClient in a test context.
-    '''
-
-    set_module_args(dict(
-        api_version=DEFAULT_DOCKER_API_VERSION
-    ))
-
-    with patch.object(AnsibleDockerClient, 'version') as version:
-        version.return_value = dict(
-            ApiVersion=DEFAULT_DOCKER_API_VERSION
-        )
-
-        client = AnsibleDockerClient()
-
-    return client
-
-
-@pytest.mark.parametrize("config", CREDENTIAL_HELPERS)
-def test_docker_credential_helpers(config):
-    client = get_docker_client()
-
-    try:
-        store = client.get_credential_store_instance(EXAMPLE_REGISTRY, os.path.join(test_data_dir, config))
-
-        assert type(store) == Store
-    except StoreError as e:
-        pytest.skip("Credential helper not available.")
-
-
-def test_docker_legacy_crednetials():
-    client = get_docker_client()
-
-    store = client.get_credential_store_instance(EXAMPLE_REGISTRY, os.path.join(test_data_dir, "none-config.json"))
-
-    assert type(store) == DockerFileStore
-    assert store.get(EXAMPLE_REGISTRY) == dict(
-        Username="abc",
-        Secret="123"
-    )
