+++ test/units/modules/cloud/docker/test_data/desktop-config.json
