+++ test/units/modules/cloud/docker/test_data/none-config.json
