+++ test/units/modules/cloud/docker/test_data/osxkeychain-config.json
